Separate checks for windows and nix systems. Windows has no logrotate.
