+++ file.go
@@ -8,7 +8,6 @@ import (
 	"path/filepath"
 	"strconv"
 	"sync"
-	"syscall"
 	"time"
 )
 
@@ -90,9 +89,7 @@ func (r *FileOutput) reopenChecker(exit chan interface{}) {
 	for {
 		select {
 		case <-ticker.C:
-			r.Lock()
-			r.check()
-			r.Unlock()
+			r.doWithCheck(func(){})
 		case <-exit:
 			return
 		}
@@ -112,60 +109,13 @@ func (r *FileOutput) reopen() *os.File {
 	return r.f
 }
 
-func (r *FileOutput) check() {
-	now := time.Now()
-
-	if now.Before(r.checkNext) {
-		return
-	}
-
-	r.checkNext = time.Now().Add(r.timeout)
-
-	fInfo, err := r.f.Stat()
-	if err != nil {
-		fmt.Println(err.Error())
-		return
-	}
-
-	fStat, ok := fInfo.Sys().(*syscall.Stat_t)
-	if !ok {
-		fmt.Println("Not a syscall.Stat_t")
-		return
-	}
-
-	pInfo, err := os.Stat(r.path)
-	if err != nil {
-		// file deleted (?)
-		r.reopen()
-		return
-	}
-
-	pStat, ok := pInfo.Sys().(*syscall.Stat_t)
-	if !ok {
-		fmt.Println("Not a syscall.Stat_t")
-		return
-	}
-
-	if fStat.Ino != pStat.Ino {
-		// file on disk changed
-		r.reopen()
-		return
-	}
-}
-
 func (r *FileOutput) Write(p []byte) (n int, err error) {
-	r.Lock()
-	r.check()
-	n, err = r.f.Write(p)
-	r.Unlock()
+	r.doWithCheck(func() {n, err = r.f.Write(p)})
 	return
 }
 
 func (r *FileOutput) Sync() (err error) {
-	r.Lock()
-	r.check()
-	err = r.f.Sync()
-	r.Unlock()
+	r.doWithCheck(func() {err = r.f.Sync()})
 	return
 }
