+++ file_check_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package zapwriter
+
+func (r *FileOutput) doWithCheck(f func()) {
+	f()
+}
